+++ nodes/complete-task.html
@@ -1,8 +1,8 @@
 <script type="text/javascript">
-    RED.nodes.registerType('todoist-query-api',{
+    RED.nodes.registerType('todoist-task-complete',{
         category: 'Todoist',
         icon: "icons/todoist.png",
-        color: '#E9967A', // Todoist base color #E44332
+        color: '#db4c3f',
         defaults: {
             name: {value:""},
             token: {
@@ -12,13 +12,13 @@
         inputs:1,
         outputs:1,
         label: function () {
-            return this.name ||  "Query API";
+            return this.name ||  "Complete a task";
         },
-        paletteLabel: 'Custom query',
+        paletteLabel: 'Complete a task',
     });
 </script>
 
-<script type="text/html" data-template-name="todoist-query-api">
+<script type="text/html" data-template-name="todoist-task-complete">
     <div class="form-row">
         <label for="node-input-token"
           ><i class="fa fa-key" ></i> Select Token</label
@@ -31,16 +31,15 @@
     </div>
 </script>
 
-<script type="text/markdown" data-help-name="todoist-query-api">
-    ## Custom API query
+<script type="text/markdown" data-help-name="todoist-task-complete">
+    ## Complete a task
 
-    Send any query to the Todoist API. The query is sent as a JSON payload, which looks like this:
-    
-    ```json
-    {
-        "endpoint": "your endpoint",
-        "method": "your method",
-        "data": { the data you wish to send (optional)}
-    }
+    This node allows you to complete (mark incomplete) a task in your Todoist account. It expects a payload with a task ID.
+
+    e.g.
+
     ```
+    {
+        "id": "123456789"
+    }
 </script>
