+++ nodes/create-task.html
@@ -1,8 +1,8 @@
 <script type="text/javascript">
-    RED.nodes.registerType('todoist-task-delete',{
+    RED.nodes.registerType('todoist-task-create',{
         category: 'Todoist',
         icon: "icons/todoist.png",
-        color: '#E9967A', // Todoist base color #E44332
+        color: '#db4c3f',
         defaults: {
             name: {value:""},
             token: {
@@ -12,13 +12,13 @@
         inputs:1,
         outputs:1,
         label: function () {
-            return this.name ||  "Delete a task";
+            return this.name ||  "Create Task";
         },
-        paletteLabel: 'Delete a task',
+        paletteLabel: 'Create Task',
     });
 </script>
 
-<script type="text/html" data-template-name="todoist-task-delete">
+<script type="text/html" data-template-name="todoist-task-create">
     <div class="form-row">
         <label for="node-input-token"
           ><i class="fa fa-key" ></i> Select Token</label
@@ -31,15 +31,17 @@
     </div>
 </script>
 
-<script type="text/markdown" data-help-name="todoist-task-delete">
-    ## Delete task
+<script type="text/markdown" data-help-name="todoist-task-create">
+    ## Create task
 
-    This node deletes a task in your Todoist account. It expects a payload with a task ID.
+    This node creates a task in your Todoist account. It expects a payload of data in the form of a JSON object with data that matches the format in the [Todoist REST API documentation](https://developer.todoist.com/rest/v1/#tasks).:
 
-    e.g.
-
-    ```
+    e.g. 
+    ```json
     {
-        "id": "123456789"
+        "content": "Test Task",
+        "project_id": 2203306141,
+        "priority": 1
     }
+    ```
 </script>
