+++ nodes/delete-task.html
@@ -1,8 +1,8 @@
 <script type="text/javascript">
-    RED.nodes.registerType('todoist-task-get-all',{
+    RED.nodes.registerType('todoist-task-delete',{
         category: 'Todoist',
         icon: "icons/todoist.png",
-        color: '#E9967A', // Todoist base color #E44332
+        color: '#db4c3f',
         defaults: {
             name: {value:""},
             token: {
@@ -12,13 +12,13 @@
         inputs:1,
         outputs:1,
         label: function () {
-            return this.name ||  "Get all tasks";
+            return this.name ||  "Delete a task";
         },
-        paletteLabel: 'Get all tasks',
+        paletteLabel: 'Delete a task',
     });
 </script>
 
-<script type="text/html" data-template-name="todoist-task-get-all">
+<script type="text/html" data-template-name="todoist-task-delete">
     <div class="form-row">
         <label for="node-input-token"
           ><i class="fa fa-key" ></i> Select Token</label
@@ -31,16 +31,15 @@
     </div>
 </script>
 
-<script type="text/markdown" data-help-name="todoist-task-get-all">
-    ## get task
+<script type="text/markdown" data-help-name="todoist-task-delete">
+    ## Delete task
 
-    This node gets all tasks in your Todoist account, no payload is required to get all tasks on account, but if you want to get all tasks of a specific project, you must pass `projectId` as a payload.
+    This node deletes a task in your Todoist account. It expects a payload with a task ID.
 
     e.g.
 
-    ```json
-    {
-        "projectId": "123456789"
-    }
     ```
+    {
+        "id": "123456789"
+    }
 </script>
