+++ nodes/delete-task.js
@@ -1,6 +1,6 @@
 module.exports = function(RED) {
-    var todoistQuery = require('../../lib/todoist-query');
-    function TodoistTaskReopen(config) {
+    var todoistQuery = require('../lib/todoist-query');
+    function TodoistTaskDelete(config) {
         RED.nodes.createNode(this,config);
 
         var node = this;
@@ -8,12 +8,11 @@ module.exports = function(RED) {
         var token = RED.nodes.getNode(config.token).credentials.token;
 
         node.on('input', function(msg) {
-            var data = msg.payload;
+            var id = msg.payload.id;
             var options = {
                 token,
-                endpoint: `tasks/${data.id}/reopen`,
-                method: 'POST',
-                data
+                endpoint: `tasks/${id}`,
+                method: 'DELETE'
             };
             todoistQuery(options)
                 .then(function(response) {
@@ -30,5 +29,5 @@ module.exports = function(RED) {
                 })
         });
     }
-    RED.nodes.registerType("todoist-task-reopen", TodoistTaskReopen);
+    RED.nodes.registerType("todoist-task-delete", TodoistTaskDelete);
 }
