+++ nodes/get-all-tasks.html
@@ -1,8 +1,8 @@
 <script type="text/javascript">
-    RED.nodes.registerType('todoist-task-complete',{
+    RED.nodes.registerType('todoist-task-get-all',{
         category: 'Todoist',
         icon: "icons/todoist.png",
-        color: '#E9967A', // Todoist base color #E44332
+        color: '#db4c3f', 
         defaults: {
             name: {value:""},
             token: {
@@ -12,13 +12,13 @@
         inputs:1,
         outputs:1,
         label: function () {
-            return this.name ||  "Complete a task";
+            return this.name ||  "Get all tasks";
         },
-        paletteLabel: 'Complete a task',
+        paletteLabel: 'Get all tasks',
     });
 </script>
 
-<script type="text/html" data-template-name="todoist-task-complete">
+<script type="text/html" data-template-name="todoist-task-get-all">
     <div class="form-row">
         <label for="node-input-token"
           ><i class="fa fa-key" ></i> Select Token</label
@@ -31,15 +31,16 @@
     </div>
 </script>
 
-<script type="text/markdown" data-help-name="todoist-task-complete">
-    ## Complete a task
+<script type="text/markdown" data-help-name="todoist-task-get-all">
+    ## get task
 
-    This node allows you to complete (mark incomplete) a task in your Todoist account. It expects a payload with a task ID.
+    This node gets all tasks in your Todoist account, no payload is required to get all tasks on account, but if you want to get all tasks of a specific project, you must pass `projectId` as a payload.
 
     e.g.
 
-    ```
+    ```json
     {
-        "id": "123456789"
+        "projectId": "123456789"
     }
+    ```
 </script>
