+++ nodes/query.html
@@ -1,8 +1,8 @@
 <script type="text/javascript">
-    RED.nodes.registerType('todoist-task-create',{
+    RED.nodes.registerType('todoist-query-api',{
         category: 'Todoist',
         icon: "icons/todoist.png",
-        color: '#E9967A', // Todoist base color #E44332
+        color: '#db4c3f', 
         defaults: {
             name: {value:""},
             token: {
@@ -12,13 +12,13 @@
         inputs:1,
         outputs:1,
         label: function () {
-            return this.name ||  "Create Task";
+            return this.name ||  "Query API";
         },
-        paletteLabel: 'Create Task',
+        paletteLabel: 'Custom query',
     });
 </script>
 
-<script type="text/html" data-template-name="todoist-task-create">
+<script type="text/html" data-template-name="todoist-query-api">
     <div class="form-row">
         <label for="node-input-token"
           ><i class="fa fa-key" ></i> Select Token</label
@@ -31,17 +31,16 @@
     </div>
 </script>
 
-<script type="text/markdown" data-help-name="todoist-task-create">
-    ## Create task
+<script type="text/markdown" data-help-name="todoist-query-api">
+    ## Custom API query
 
-    This node creates a task in your Todoist account. It expects a payload of data in the form of a JSON object with data that matches the format in the [Todoist REST API documentation](https://developer.todoist.com/rest/v1/#tasks).:
-
-    e.g. 
+    Send any query to the Todoist API. The query is sent as a JSON payload, which looks like this:
+    
     ```json
     {
-        "content": "Test Task",
-        "project_id": 2203306141,
-        "priority": 1
+        "endpoint": "your endpoint",
+        "method": "your method",
+        "data": { the data you wish to send (optional)}
     }
     ```
 </script>
