+++ nodes/reopen-task.html
@@ -1,8 +1,8 @@
 <script type="text/javascript">
-    RED.nodes.registerType('todoist-task-update',{
+    RED.nodes.registerType('todoist-task-reopen',{
         category: 'Todoist',
         icon: "icons/todoist.png",
-        color: '#E9967A', // Todoist base color #E44332
+        color: '#db4c3f',
         defaults: {
             name: {value:""},
             token: {
@@ -12,13 +12,13 @@
         inputs:1,
         outputs:1,
         label: function () {
-            return this.name ||  "Update a task";
+            return this.name ||  "Reopen a task";
         },
-        paletteLabel: 'Update a task',
+        paletteLabel: 'Reopen a task',
     });
 </script>
 
-<script type="text/html" data-template-name="todoist-task-update">
+<script type="text/html" data-template-name="todoist-task-reopen">
     <div class="form-row">
         <label for="node-input-token"
           ><i class="fa fa-key" ></i> Select Token</label
@@ -31,20 +31,15 @@
     </div>
 </script>
 
-<script type="text/markdown" data-help-name="todoist-task-update">
-    ## Update task
+<script type="text/markdown" data-help-name="todoist-task-reopen">
+    ## Re-open a task
 
-    This node updates a task in your Todoist account. It expects a payload with the task id and the updated data:
+    This node allows you to Re-open (mark incomplete) a task in your Todoist account. It expects a payload with a task ID.
 
     e.g.
 
-    ```json
-    {
-        "id": 2995104339,
-        "content": "Test Task (updated)",
-        "project_id": 2203306141,
-        "priority": 1,
-        "assignee": 2671142
-    }
     ```
+    {
+        "id": "123456789"
+    }
 </script>
