+++ package.json
@@ -4,15 +4,15 @@
     "description": "Todoist API interaction nodes for Node-RED",
     "node-red": {
         "nodes": {
-            "add-token": "./nodes/config/index.js",
-            "todoist-task-get": "./nodes/get-task/index.js",
-            "todoist-task-get-all": "./nodes/get-all-tasks/index.js",
-            "todoist-task-create": "./nodes/create-task/index.js",
-            "todoist-task-update": "./nodes/update-task/index.js",
-            "todoist-task-complete": "./nodes/complete-task/index.js",
-            "todoist-task-reopen": "./nodes/reopen-task/index.js",
-            "todoist-task-delete": "./nodes/delete-task/index.js",
-            "todoist-query-api": "./nodes/query/index.js"
+            "add-token": "./nodes/config.js",
+            "todoist-task-get": "./nodes/get-task.js",
+            "todoist-task-get-all": "./nodes/get-all-tasks.js",
+            "todoist-task-create": "./nodes/create-task.js",
+            "todoist-task-update": "./nodes/update-task.js",
+            "todoist-task-complete": "./nodes/complete-task.js",
+            "todoist-task-reopen": "./nodes/reopen-task.js",
+            "todoist-task-delete": "./nodes/delete-task.js",
+            "todoist-query-api": "./nodes/query.js"
         }
     },
     "license": "MIT",
